fix(budget): true dynamic calculation of safe-to-spend and remaining free money, removed onboarding, change to rolling pay cycle dates

+++ backend/app/schemas/anti_impulse.py
@@ -25,7 +25,7 @@ class AntiImpulseResolveRequest(BaseModel):
 
 class AntiImpulseSessionRead(BaseModel):
     id: uuid.UUID
-    planned_purchase_amount: Decimal
+    planned_purchase_amount: float
     category: str
     risk_score: int
     cooldown_seconds: int
